Move mock_app_version fixture to unit tests module
